lesson8/requirements.txt
allure-pytest
pytest
pytest-xdist
requests
selenium